refactor(mentor): DTO 패키지 구조 변경
- response와 prequest 분리
