+++ src/main/java/com/sprint/mission/discodeit/dto2/request/MessageCreateRequest.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto2;
+package com.sprint.mission.discodeit.dto2.request;
 
 import java.util.List;
 import java.util.UUID;
