+++ src/main/java/com/sprint/mission/discodeit/dto2/request/ReadStatusCreateRequest.java
@@ -1,10 +1,9 @@
-package com.sprint.mission.discodeit.dto2;
+package com.sprint.mission.discodeit.dto2.request;
 
 import java.time.Instant;
 import java.util.UUID;
 
-public record ReadStatusUpdateRequest(
-        UUID readStatusId,
+public record ReadStatusCreateRequest(
         UUID userId,
         UUID channelId,
         Instant lastReadAt
