+++ src/main/java/com/sprint/mission/discodeit/dto2/request/ReadStatusUpdateRequest.java
@@ -1,9 +1,10 @@
-package com.sprint.mission.discodeit.dto2;
+package com.sprint.mission.discodeit.dto2.request;
 
 import java.time.Instant;
 import java.util.UUID;
 
-public record ReadStatusCreateRequest(
+public record ReadStatusUpdateRequest(
+        UUID readStatusId,
         UUID userId,
         UUID channelId,
         Instant lastReadAt
