+++ src/main/java/com/sprint/mission/discodeit/dto2/request/UserCreateRequest.java
@@ -1,6 +1,4 @@
-package com.sprint.mission.discodeit.dto2;
-
-import com.sprint.mission.discodeit.entity.UserStatus;
+package com.sprint.mission.discodeit.dto2.request;
 
 public record UserCreateRequest(
         String username,
