+++ src/main/java/com/sprint/mission/discodeit/dto2/request/UserStatusCreateRequest.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto2;
+package com.sprint.mission.discodeit.dto2.request;
 
 import java.time.Instant;
 import java.util.UUID;
