+++ src/main/java/com/sprint/mission/discodeit/dto2/request/UserUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto2;
+package com.sprint.mission.discodeit.dto2.request;
 
 import java.util.UUID;
