+++ src/main/java/com/sprint/mission/discodeit/dto2/response/ChannelResponse.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto2;
+package com.sprint.mission.discodeit.dto2.response;
 
 import com.sprint.mission.discodeit.entity.ChannelType;
