+++ src/main/java/com/sprint/mission/discodeit/service/ReadStatusService.java
@@ -1,7 +1,7 @@
 package com.sprint.mission.discodeit.service;
 
-import com.sprint.mission.discodeit.dto2.ReadStatusCreateRequest;
-import com.sprint.mission.discodeit.dto2.ReadStatusUpdateRequest;
+import com.sprint.mission.discodeit.dto2.request.ReadStatusCreateRequest;
+import com.sprint.mission.discodeit.dto2.request.ReadStatusUpdateRequest;
 import com.sprint.mission.discodeit.entity.ReadStatus;
 
 import java.util.List;
